pyproject.toml
[tool.poetry]
name = "cw-7"
version = "0.1.0"
description = ""
authors = ["Victor Volkov <vektorn1212@gmail.com>"]
readme = "README.md"
packages = [{include = "cw_7"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.26.5"
coverage = "^7.3.2"
celery = "^5.3.6"
requests = "^2.31.0"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
